Soduku-Solver
It utilizes arrays to implement the well-known Snake game, where a user-controlled snake moves around a grid, consuming food to grow while avoiding collisions with walls and its own body. This project not only offers an enjoyable gaming experience but also demonstrates the application of data structures and algorithms in game development.